requirements.txt
fastapi
uvicorn
pandas
numpy
yfinance
python-multipart
google-generativeai
python-dotenv
